+++ docs/.vuepress/components/Input/demo-clearable.vue
@@ -4,34 +4,28 @@
  * @Author: langxue
  * @Date: 2019-11-12 11:07:52
  * @LastEditors: langxue
- * @LastEditTime: 2019-11-27 15:05:31
+ * @LastEditTime: 2019-12-04 14:04:59
  -->
 <template>
   <div id="input-demo">
     <z-input type="text"
-        v-model="value"
         clearable
-        ref="input"
-        @clear="clear"
-        placeholder="请输入内容" />
-        {{value}}
+        v-model="value"
+          placeholder="请输入内容"
+          @clear="handleClear" />
   </div>
 </template>
 <script>
 export default {
   data() {
     return {
-      value: 0
+      value: ''
     };
   },
-  mounted() {
-    this.$refs.input.focus();
-  },
   methods: {
-    clear(e) {
-      this.value = null
-      console.log('clear');
-    },
+    handleClear(e) {
+      console.log('清空内容', e);
+    }
   }
 };
 </script>
