+++ docs/.vuepress/components/Input/demo-disabled.vue
@@ -4,13 +4,13 @@
  * @Author: 
  * @Date: 2019-11-14 16:03:05
  * @LastEditors: langxue
- * @LastEditTime: 2019-11-14 17:07:58
+ * @LastEditTime: 2019-12-04 14:02:47
  -->
 <template>
   <div id="input-demo-disabled">
     <z-input type="text"
-          placeholder="请输入内容"
-          disabled></z-input>
+        disabled
+          placeholder="请输入内容" />
   </div>
 </template>
 <style lang="stylus">
