+++ docs/.vuepress/components/Input/demo-icon.vue
@@ -4,31 +4,30 @@
  * @Author: 
  * @Date: 2019-11-14 16:03:05
  * @LastEditors: langxue
- * @LastEditTime: 2019-11-22 10:23:33
+ * @LastEditTime: 2019-12-04 14:20:47
  -->
 <template>
   <div id="input-demo-icon">
     <div class="attr">
       <strong>属性方式：</strong>
-      <z-input type="text"
-        placeholder="请输入内容"
-        suffixIcon="zxclose"
-        @click="handleClick"
-        class="item-input" />
         <z-input type="text"
-        placeholder="请输入内容"
+        suffixIcon="zxsousuo"
+          placeholder="后置icon"
+          class="item-input" />
+        <z-input type="text"
+        placeholder="前置icon"
         prefixIcon="zxsousuo"
         class="item-input" />
     </div>
     <div class="attr">
       <strong>slot方式：</strong>
       <z-input type="text"
-        placeholder="请输入内容"
+        placeholder="slot 后置icon"
         class="item-input">
-        <i slot="suffix" class="iconfont zxclose"></i>
+        <i slot="suffix" class="iconfont zxsousuo"></i>
       </z-input>
-        <z-input type="text"
-        placeholder="请输入内容"
+      <z-input type="text"
+        placeholder="slot 前置icon"
         class="item-input">
         <i slot="prefix" class="iconfont zxsousuo"></i>
       </z-input>
