+++ docs/vue/Input/README.md
@@ -4,7 +4,7 @@
  * @Author: bhabgs
  * @Date: 2019-11-12 09:06:17
  * @LastEditors: langxue
- * @LastEditTime: 2019-11-29 13:41:21
+ * @LastEditTime: 2019-12-04 14:06:46
  -->
 
 # input 输入
@@ -31,19 +31,19 @@
 
 :::
 
-::: explain 标签
+::: explain 可清空
 
-<templatePc name="Input-demo-label">
+<templatePc name="Input-demo-clearable">
 
-<<< @/docs/.vuepress/components/Input/demo-label.vue
-
-### 可清空
-
-<templates-demo name="Input-demo-clearable">
 <<< @/docs/.vuepress/components/Input/demo-clearable.vue
-</templates-demo>
 
-### 尺寸
+</templatePc>
+
+:::
+
+::: explain 尺寸
+
+<templatePc name="Input-demo-size">
 
 <<< @/docs/.vuepress/components/Input/demo-size.vue
 
@@ -90,14 +90,14 @@
 | type           | 文本框类型           | String          | false    | text 和其他原生 input 的 type 值 | text       |
 | placeholder    | 文本框占位文本       | String          | false    | -                                | 请输入内容 |
 | disabled       | 禁用                 | Boolean         | false    | -                                | false      |
-| label          | 文本框标签文本       | String          | false    | -                                | -          |
-| labelWidth     | 文本框标签文本宽度   | String          | false    | -                                | auto       |
+<!-- | label          | 文本框标签文本       | String          | false    | -                                | -          |
+| labelWidth     | 文本框标签文本宽度   | String          | false    | -                                | auto       | -->
 | size           | 文本框尺寸           | String          | false    | medium，small，mini              | medium     |
-| min            | 文本框输入最小值     | Number          | false    | -                                | -          |
-| max            | 文本框输入最大值     | Number          | false    | -                                | -          |
+<!-- | min            | 文本框输入最小值     | Number          | false    | -                                | -          |
+| max            | 文本框输入最大值     | Number          | false    | -                                | -          | -->
 | suffixIcon     | 后置 icon            | String          | false    | -                                | -          |
 | prefixIcon     | 前置 icon            | String          | false    | -                                | -          |
-| showWordLimit  | 是否显示输入字数统计 | Boolean         | false    | -                                | false      |
+<!-- | showWordLimit  | 是否显示输入字数统计 | Boolean         | false    | -                                | false      | -->
 | value(v-model) | 绑定值               | string / number | false    | -                                | -          |
 
 :::
@@ -118,5 +118,5 @@
 | blur     | 在 Input 失去焦点时触发             | -    | -        | -      | -      |
 | input    | 在 Input 值改变时触发               | -    | -        | -      | -      |
 | change   | 在 Input 失去焦点且值发生改变时触发 | -    | -        | -      | -      |
-
+| clear   | 清空input内容时触发 | -    | -        | -      | -      |
 :::
