+++ src/components/Input/index.tsx
@@ -1,268 +1,125 @@
 import { Component, Vue, Prop, Model, Watch } from 'vue-property-decorator';
-import {VNode} from 'vue';
+import inputProps from './inputProps';
+
+const baseInput = Vue.extend({
+  props: {...inputProps},
+  model: {
+    prop: 'value',
+    event: 'change.value',
+  },
+  data() {
+    const { value, modelValue } = this.$props;
+    return {
+      stateValue: modelValue ? modelValue : value,
+    }
+  },
+})
 
 @Component
-export default class ZInput extends Vue {
-  @Model('input', {
-    type: [Number, String]
-  }) public inputValue!: number | string;
+export default class ZInput extends baseInput {
 
-  // value
-  @Prop({
-    type: [Number, String]
-  }) public value!: number | string;
-
-  // 类型
-  @Prop({
-    default: 'text',
-    type: String
-  }) public type!: string;
-
-  // 占位符
-  @Prop({
-    default: '请输入内容',
-    type: String
-  }) public placeholder!: string;
-
-  // 禁用
-  @Prop({
-    default: false,
-    type: Boolean
-  }) public disabled!: boolean;
-
-  // label
-  @Prop({
-    default: '',
-    type: String
-  }) public label!: string;
-
-  // label-width
-  @Prop({
-    default: 'auto',
-    type: String
-  }) public labelWidth!: string;
-  
-  // size
-  @Prop({
-    default: '',
-    type: String
-  }) public size!: string;
-
-  // 输入最小限制
-  @Prop({
-    default: 0,
-    type: Number
-  }) public min!: number;
-
-  // 输入最大限制
-  @Prop({
-    type: Number
-  }) public max!: number;
-
-  // 是否显示输入字符
-  @Prop({
-    type: Boolean
-  }) public showWordLimit!: boolean;
-
-  // 后置icon
-  @Prop({
-    type: String
-  }) public suffixIcon!: string;
-
-  // 前置icon
-  @Prop({
-    type: String
-  }) public prefixIcon!: string;
-
-  // 看清空
-  @Prop({
-    type: Boolean
-  }) public clearable!: boolean;
-
-  @Watch('value') public getValue(val) {
-    console.log(val);
+  public mounted() {
+    // console.log(this.$props);
   }
 
-  public hovering: boolean = false;
-  public focused: boolean = false;
-
-  // 设置input框的类型
-  public setInputType() {
-    return this.type;
-  }
-
-  // 设置placeholder
-  public setPlaeholder() {
-    return this.placeholder;
-  }
-
-  // 设置禁用状态
-  public setDisabled() {
-    return this.disabled;
-  }
-
-  // 设置最小限制和最大限制
-  public setLength() {
-    const { min, max } = this;
-    return {
-      min,
-      max
-    };
-  }
-
-  // 设置是否显示输入字符
-  public setShowWordLimit() {
-    const { max } = this;
-    const wordLimitClass = {
-      'z-input-show-limit': this.showWordLimit
-    };
-    const wordLimit = <span class={ wordLimitClass }>2/{ max }</span>;
-    return wordLimit;
-  }
-
-  // 设置class
+  // set input classnames
   public setInputClass() {
-    const { disabled, size, suffixIcon, prefixIcon, $slots, clearable } = this;
+    const { $slots, $props } = this;
     return {
-      [`z-input-${size}`]: size === 'medium' || '' ? null : size,
-      'z-input-disabled': disabled,
-      'z-input-suffix': suffixIcon || $slots.suffix || clearable,
-      'z-input-prefix': prefixIcon || $slots.prefix
+      [`z-input-${$props.size}`]: $props.size == 'default' ? '' : $props.size,
+      'z-input-disabled': $props.disabled,
+      'z-input-suffix': $props.suffixIcon || $slots.suffix || $props.clearable,
+      'z-input-prefix': $props.prefixIcon || $slots.prefix
     };
   }
 
-  // handleBlur
-  public handleBlur(e: any) {
-    this.$emit('blur', e.target.value);
+  // set icon classnames
+  public setIconClass() {
+    const { $props, $slots } = this;
+    const prefix = $props.prefixIcon || $slots.prefix;
+    return {
+      [`z-input-icon-${$props.size}`]: prefix ? $props.size : '',
+    }
+  }
+
+  // render prefixIcon or slot prefixIcon
+  public renderPrefixIcon(): JSX.Element {
+    const { $props, $slots } = this;
+    const prefixIcon = $props.prefixIcon || $slots.prefix;
+    if (prefixIcon) {
+      return <span class={["z-input-prefix-icon"]}>
+        <i class={["iconfont z-icon", $props.prefixIcon]}></i>
+        {$slots.prefix}
+      </span>
+    }
+    return null;
+  }
+
+  // render clear icon
+  public renderClearIcon() {
+    const { clearable, disabled } = this.$props;
+    const { stateValue, handleClear } = this;
+    // 不清空，禁用，值为空
+    if (!clearable || disabled || stateValue === '' || stateValue === null || stateValue === undefined) {
+      return null;
+    }
+    const clearIcon = <i class="iconfont zxclose z-icon z-input-clear-icon" onClick={handleClear}></i>;
+    return clearIcon;
+  }
+
+  // suffix
+  public renderSuffixIcon() {
+    const { $props, $slots } = this;
+    // $slots 直接渲染 $slots.suffix 否则 $props.suffixIcon
+    // return $slots.suffix;
+    return $slots.suffix ? $slots.suffix : <i class={["iconfont z-icon", $props.suffixIcon]}></i>
+  }
+
+  // render suffixIcon
+  public renderSuffix() {
+    const { $props, $slots, renderClearIcon, renderSuffixIcon } = this;
+    const suffixClass = {
+      "z-input-suffix-icon": true,
+      "z-input-clear-icon": $props.clearable,
+    }
+    return <span class={suffixClass}>
+        {renderClearIcon()}
+        {renderSuffixIcon()}
+      </span>
   }
 
   // handleInput
   public handleInput(e: any) {
-    this.$emit('input', e.target.value);
+    const value = e.target.value;
+    this.$emit('input', value);
+    this.stateValue = value;
   }
 
-  // handleChange
-  public handleChange(e: any) {
-    this.$emit('change', e.target.value);
-  }
-
-  // handleFocus
-  public handleFocus(e: any) {
-    this.focused = true;
-    this.$emit('focus', event);
-  }
-
-  public focus() {
-    const inputElement: any = this.$refs.input;
-    inputElement.focus();
-  }
-
-  // handleClickIcon
-  public handleClickIcon(e: any) {
-    e.stopPropagation();
-    this.$emit('click', e);
-  }
-
-  // 鼠标进入
-  public handleMouseEnter(e: any) {
-    e.stopPropagation();
-    this.hovering = true;
-  }
-
-  // 鼠标离开
-  public handleMouseLeave(e: any) {
-    this.hovering = false;
-  }
-
-  // 清空
+  // handleClear
   public handleClear(e: any) {
-    e.stopPropagation();
-    this.$emit('input', '');
-    this.$emit('change', '');
-    this.$emit('clear');
-  }
-
-  // render Label
-  public renderLabel() {
-    const { label, labelWidth } = this;
-    const labelWidthStyle = {
-      width: labelWidth
-    };
-    return label ? <span class={'z-input-label'} style={labelWidthStyle}>{label}:</span> : null;
-  }
-
-  // render prefixIcon or slot prefixIcon
-  public renderPrefixIcon() {
-    const { prefixIcon, $slots } = this;
-    const prefixRender = prefixIcon || $slots.prefix ? <span class="z-input-prefix-icon">
-      <span class="z-input-prefix-icon-inner">
-        <i class={['iconfont', prefixIcon]} if="prefixIcon"></i>
-        {$slots.prefix}
-      </span>
-    </span> : null;
-    return prefixRender;
-  }
-
-  // render suffixIcon
-  public renderSuffixIcon() {
-    const { suffixIcon, $slots, handleClickIcon, showClear, handleClear } = this;
-    if (showClear()) {
-      return <span class="z-input-suffix-icon">
-        <span class="z-input-suffix-icon-inner">
-          <i class='iconfont zxclose' on-mousedown={(e) => e.stopPropagation()} on-click={handleClear}></i>
-          {$slots.suffix}
-        </span>
-      </span>
-    } else {
-      const suffix = suffixIcon || $slots.suffix ? <span class="z-input-suffix-icon" on-click={handleClickIcon}>
-        <span class="z-input-suffix-icon-inner">
-          <i class={['iconfont', suffixIcon]} if="suffixIcon"></i>
-          {$slots.suffix}
-        </span>
-      </span> : null;
-      return suffix;
-    }
-  }
-
-  // 显示清空图标
-  public showClear() {
-    return !this.disabled && this.hovering && this.clearable;
-  }
-
-  // 是否显示后置内容
-  public getVisibleSuffix() {
-    const {$slots, suffixIcon, showClear} = this;
-    return $slots.suffix || suffixIcon || showClear()
+    this.stateValue = '';
+    this.$emit('clear', e);
   }
 
   // render input
-  public renderInput() {
-    const { setInputType, setPlaeholder, setDisabled, setLength, handleBlur, handleInput, handleChange, handleFocus } = this;
-    const limitLength = setLength();
+  public renderInput(): JSX.Element {
+    const { stateValue, $props, handleInput } = this;
     return <input ref="input"
-      type={setInputType()}
-      placeholder={setPlaeholder()}
-      class='z-input-inner'
-      disabled={setDisabled()}
-      maxLength={limitLength.max}
-      minLength={limitLength.min}
-      on-blur={handleBlur}
-      on-input={handleInput}
-      on-change={handleChange}
-      on-focus={handleFocus} />;
+      value= {stateValue}
+      disabled={$props.disabled}
+      class={['z-input-inner']}
+      placeholder={$props.placeholder}
+      on-input={handleInput} />;
   }
 
   public render() {
-    const { renderInput, renderLabel, setInputClass, renderSuffixIcon, renderPrefixIcon, handleMouseEnter, handleMouseLeave } = this;
+    const { renderInput, renderPrefixIcon, renderSuffix, setInputClass } = this;
     return (
-      <div class={['z-input', setInputClass()]} on-mouseenter={handleMouseEnter}
-      on-mouseleave={handleMouseLeave} >
+      <div class={["z-input", setInputClass()]}>
         {renderPrefixIcon()}
-        {renderLabel()}
         {renderInput()}
-        {renderSuffixIcon()}
-        {/* {setShowWordLimit()} */}
+        {renderSuffix()}
       </div>
-    );
+    )
   }
 }
